simplify repository api for review
